# Golangci lint v2 WIP