lint for v2