@jidicula 2021-04-03
`internal/imported_tests/unmarshal_imported_test.go` has a lot of unused variables that `golangci-lint` is flagging. Is it ok if I nuke them or are they going to be used as the v2 WIP continues?