@vincentserpoul 2021-04-04
@pelletier I think regular merges/checks would help a lot to avoid a big painful rebase if you continue development