@vincentserpoul 2021-04-05
> FYI I ran the benchmarks between the base of this branch and tip, and benchstat is reporting some regression on the UnmarshalSimple test:
> 
> ```
> name                  old time/op    new time/op    delta
> UnmarshalSimple/v2-8    1.02µs ± 3%    1.04µs ± 2%  +1.78%  (p=0.012 n=10+10)
> ReferenceFile/v2-8      34.3µs ± 4%    35.0µs ± 5%    ~     (p=0.190 n=10+10)
> 
> name                  old speed      new speed      delta
> ReferenceFile/v2-8     153MB/s ± 4%   150MB/s ± 5%    ~     (p=0.190 n=10+10)
> 
> name                  old alloc/op   new alloc/op   delta
> ReferenceFile/v2-8      11.8kB ± 0%    11.8kB ± 0%    ~     (all equal)
> 
> name                  old allocs/op  new allocs/op  delta
> ReferenceFile/v2-8         253 ± 0%       253 ± 0%    ~     (all equal)
> ```
> 
> (using `go test '-bench=.*/v2' -count=10`)
> 
> Not sure what would be causing this though.

a few hundreds nanosecs, is it systematic? Could it be your device having a different background noise?