@vincentserpoul 2021-04-05
Let's recap here the conclusions of these lints:
* New errors or const should not be made public
* const should be declared as close as possible as their use
* Revert all autofixes
* autofix to be done one file by one file, to be sure it's not doing dumb things
* do not use switch when it brings a label (to be confirmed)
* wsl to be deactvated (to be confirmed)
*