@vincentserpoul 2021-04-05
> > @jidicula unused variables might be referenced in commented-out code in those tests. They are commented out because the API for those tests isn't there yet. In those cases, I think the variables should be commented out too. However I suspect there may be some legit unused variables in those tests.
> 
> Ok, I'll comment them out instead of removing them so the linter can be silenced _for now_. I'll mark them with TODO as well for further consideration when the API has been built out.


you can also use //nolint:deadcode