@pelletier 2021-04-05
> a few hundreds nanosecs, is it systematic? Could it be your device having a different background noise?

Re-ran it and it looked better:

```
$ benchstat bench_origin_v2.txt bench_golangci_v2.txt         
name                  old time/op    new time/op    delta
UnmarshalSimple/v2-8    1.02µs ± 3%    1.03µs ± 4%   ~     (p=0.671 n=10+10)
ReferenceFile/v2-8      34.3µs ± 4%    34.9µs ± 5%   ~     (p=0.218 n=10+10)

name                  old speed      new speed      delta
ReferenceFile/v2-8     153MB/s ± 4%   150MB/s ± 5%   ~     (p=0.218 n=10+10)

name                  old alloc/op   new alloc/op   delta
ReferenceFile/v2-8      11.8kB ± 0%    11.8kB ± 0%   ~     (all equal)

name                  old allocs/op  new allocs/op  delta
ReferenceFile/v2-8         253 ± 0%       253 ± 0%   ~     (all equal)
```

I was just flagging it to make sure we don't trigger something like unexpected escaping with the refactoring.