@vincentserpoul 2021-04-07
I rebased on @jidicula merge, kept my previous mods on a separate branch and restarted on this one, added the conf for wsl (but there are still added spaces) and took into account your remarks about errors and const.

Let me know if that's ok for you, before I continue on the rest @pelletier