@vincentserpoul 2021-04-07
it would be great to merge indeed, bit by bit so you can also have the config on your local, to be able to enable golangci-lint when you dev on your side as well!